pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "video-frame-viewer"
version = "0.2.6"
description = "EEG/EOG/EAR video frame viewer and annotator."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Video Frame Viewer Maintainers" }]
license = { text = "Proprietary" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Operating System :: OS Independent",
  "License :: Other/Proprietary License",
]
dependencies = [
  "PyQt5>=5.15.10,<5.16",
  "opencv-python-headless>=4.9.0,<4.11.0",
  "numpy>=1.23,<2.1",
  "pyqtgraph>=0.13,<0.14",
  "mne>=1.6,<1.7",
  "platformdirs>=4.2,<5.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.0,<0.6.0",
  "pytest>=7.4,<9.0",
]

[project.scripts]
drowsy-viewer = "video_frame_viewer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/video_frame_viewer"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
